requirements.txt
Flask
requests
google-generativeai
gunicorn
flask-cors